small fixes after review
