+++ src/components/app-header/app-header.module.css
@@ -4,6 +4,7 @@
   flex-direction: row;
   flex-wrap: wrap;
 }
+
 .menu{
   flex: 5;
   display: flex;
@@ -12,6 +13,7 @@
   gap: 25px;
   justify-content: center;
 }
+
 .logo{
   margin: auto;
   flex: 3;
