+++ src/components/burger-constructor/burger-constructor.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import PropTypes from 'prop-types'
-import {ingredientPropTypes} from '../../utils/dataPropTypes'
+import PropTypes from 'prop-types';
+import {ingredientPropTypes} from '../../utils/dataPropTypes';
 
 import {
   ConstructorElement,
