+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -2,10 +2,12 @@
   margin: 40px auto 20px 0;
   font-size: 36px;
 }
+
 .h3{
   margin: 24px auto 8px 0;
   font-size: 24px;
 }
+
 .ingredientsContainer{
   display: flex;
   flex-direction: row;
