+++ src/components/burger-ingredients/ingredient-card/ingredient-card.jsx
@@ -1,4 +1,3 @@
-// import HeadButton from '../head-button/head-button';
 import styles from './ingredient-card.module.css'
 import {
   Counter,
@@ -9,13 +8,13 @@ import React from 'react';
 export default function IngredientCard ({ingredient}) {
   return (
     <li key={ingredient._id} className={styles.card}>
-      <Counter count={1} size="default" />
+      <Counter count={1} size='default' />
       <img className={styles.mainImage} src={ingredient.image_large} alt={ingredient.name}/>
       <div className={styles.price}>
         <span>
           {ingredient.price}
         </span>
-        <CurrencyIcon type="primary" />
+        <CurrencyIcon type='primary' />
       </div>
       <p className={styles.name}>
         {ingredient.name}
